+++ demo/desktop-starter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # mandu-desktop-starter
 
+## 0.1.62
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mandujs/core@0.54.12
+
 ## 0.1.61
 
 ### Patch Changes